Substitute Cal.com with Google Calendar appointment schedule (#206)
I saw that Google Calendar now offers this functionality, without the
need for another service we need to share our data and calendar with.

This is how it looks like for me:

https://calendar.app.google/177nRPdJV4AtkWQh6

What to you think about getting rid of Cal.com and moving to Google
Calendar instead?

## Checklist

- [ ] I have checked the Deploy Preview link in the netlify bot's
comment, and my changes look good
- [ ] I tested my contribution locally, and it seems to work fine.
- [ ] Code and workflow changes are sufficiently documented.
- [ ] If new pages are added, maintainers are assigned for that
document.
- [ ] If the document falls under a controlled category, a controlled
document banner is present.