Add dependency management page (#298)
As requested by Sid
[here](https://github.com/open-energy-transition/oet-zib-manual/pull/12#issuecomment-3570716105).

Supersedes some content in #269 

I haven't filled in the `uv` config recommendation section as I don't
use it so don't know what would be recommended.

I would be inclined to recommend `pixi` for all but pure Python
packages, but realise that several use `uv` already and so didn't want
to push for it too hard. The reason being that those working on Linux
often miss the fact that they are utilising non-python libs that others
will need. `pixi` allows those dependencies to be explicitly referenced
in the project dependencies for better cross-platform compatibility.

## Changes Proposed in This Pull Request

## Checklist

- [ ] I have checked the Deploy Preview link in the netlify bot's
comment, and my changes look good
- [ ] I tested my contribution locally, and it seems to work fine.
- [ ] Code and workflow changes are sufficiently documented.
- [ ] If new pages are added, maintainers are assigned for that
document.
- [ ] If the document falls under a controlled category, a controlled
document banner is present.

---------

Co-authored-by: Siddharth Krishna <siddharth-krishna@users.noreply.github.com>